# shared/openapi/power-meter-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "CredentialRotationRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "maxLength": 100,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "typed_confirmation": {
            "const": "ROTATE SENSOR CREDENTIALS",
            "title": "Typed Confirmation",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key",
          "typed_confirmation"
        ],
        "title": "CredentialRotationRequest",
        "type": "object"
      },
      "DeviceEnrollmentRequest": {
        "additionalProperties": false,
        "properties": {
          "enrollment_token": {
            "maxLength": 256,
            "minLength": 32,
            "title": "Enrollment Token",
            "type": "string"
          },
          "firmware_version": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Firmware Version",
            "type": "string"
          },
          "hardware_fingerprint": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Hardware Fingerprint",
            "type": "string"
          },
          "protocol_id": {
            "const": "pm-protocol/1.0.0",
            "title": "Protocol Id",
            "type": "string"
          }
        },
        "required": [
          "enrollment_token",
          "protocol_id",
          "firmware_version",
          "hardware_fingerprint"
        ],
        "title": "DeviceEnrollmentRequest",
        "type": "object"
      },
...
      "DeviceUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "friendly_name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Friendly Name"
          },
          "measurement_scope": {
            "anyOf": [
              {
                "enum": [
                  "energy_only",
                  "allocated_account",
                  "full_account"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurement Scope"
          },
          "measurement_scope_confirmation": {
            "anyOf": [
              {
                "enum": [
                  "I VERIFIED THIS METER COVERS THE FULL ACCOUNT",
                  "I VERIFIED THIS ALLOCATION SCOPE"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurement Scope Confirmation"
          }
        },
        "title": "DeviceUpdateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PowerMeter V2 API",
    "version": "0.1.0-rc.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/devices": {
      "get": {
        "operationId": "list_devices_api_v1_devices_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "pm_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pm Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Devices Api V1 Devices Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Devices",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/devices/enroll": {
      "post": {
        "operationId": "enroll_device_api_v1_devices_enroll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceEnrollmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Enroll Device Api V1 Devices Enroll Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enroll Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}": {
      "patch": {
        "operationId": "update_device_api_v1_devices__device_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "pm_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pm Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Device Api V1 Devices  Device Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Device",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/devices/{device_id}/credentials/rotate": {
      "post": {
        "operationId": "rotate_device_credential_api_v1_devices__device_id__credentials_rotate_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "pm_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pm Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRotationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Rotate Device Credential Api V1 Devices  Device Id  Credentials Rotate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotate Device Credential",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}/credentials/rotations/{rotation_id}/cancel": {
      "post": {
        "operationId": "cancel_device_credential_rotation_api_v1_devices__device_id__credentials_rotations__rotation_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "rotation_id",
            "required": true,
            "schema": {
              "title": "Rotation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "pm_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pm Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRotationCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cancel Device Credential Rotation Api V1 Devices  Device Id  Credentials Rotations  Rotation Id  Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Device Credential Rotation",
        "tags": [
          "devices"
        ]
      }
    },
...
    "/api/v1/enrollment-tokens": {
      "post": {
        "operationId": "create_enrollment_token_api_v1_enrollment_tokens_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "pm_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pm Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollmentTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Enrollment Token Api V1 Enrollment Tokens Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Enrollment Token",
        "tags": [
          "devices"
        ]
      }
    },
...
  }
}